+++ packages/expo-splash-screen/package.json
@@ -5,10 +5,7 @@
   "main": "build/index.js",
   "types": "build/index.d.ts",
   "exports": {
-    ".": {
-      "react-native": "./build/index.native.js",
-      "default": "./build/index.js"
-    },
+    ".": "./build/index.js",
     "./plugin": "./plugin/build/index.js",
     "./package.json": "./package.json",
     "./app.plugin.js": "./app.plugin.js"
