+++ packages/expo-splash-screen/src/SplashScreen.native.ts
